Update .phpcs.xml issues and fix an undeclared variable
I was wondering about the undeclared variable. According to the Git
history this was there from the beginning.

I could think of multiple ways to fix it. I think the current behavior
is that Title::newFromDBkey() just makes this code crash with an
exception, because it can't accept null. One fix is to just continue
the loop. After reading
https://www.mediawiki.org/wiki/Extension:NamespacePopups
I believe the idea is that if the namespace is mapped to "*", the
original namespace of the link should be reused. This is what this
patch does.

Change-Id: Ibce6de64841f7c8791f0abbca05a7351139d78c3